nios2: Explicitely cast parameters to __access_ok
On Nios2 NOMMU __access_ok is an inline function taking two unsigned
long as its parameters, so cast the values passed to it to prevent the
following warning:

warning: passing arg 1 of `__access_ok' makes integer from pointer without a cast

+++ arch/nios2/include/asm/uaccess.h
@@ -79,7 +79,7 @@ static inline int __access_ok(unsigned long addr, unsigned long size)
 #endif /* CONFIG_MMU */
 
 #define access_ok(type, addr, len)		\
-	likely(__access_ok((addr), (len)))
+	likely(__access_ok((unsigned long)(addr), (unsigned long)(len)))
 
 #ifdef CONFIG_MMU
 extern long __copy_from_user(void* to, const void __user *from, unsigned long n);
